Decompile func0f110c5c

+++ src/include/game/data/data_0083d0.h
@@ -83,7 +83,7 @@ extern u32 var800628d4;
 extern u32 var80062914;
 extern s32 g_NumPlayers;
 extern u8 var80062940;
-extern u32 var80062944;
+extern u8 var80062944;
 extern u32 var80062948;
 extern u32 var80062950;
 extern bool g_WeatherTickEnabled;
